web/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
apscheduler==3.10.4
pydantic==2.5.3
python-dotenv==1.0.0
aiohttp==3.9.1
pyyaml==6.0.1
boto3>=1.34.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
